requirements.txt
Django == 3.1.7
django-filter == 2.4.0
djangorestframework == 3.12.4
psycopg2-binary == 2.8.6
pytest == 6.2.3
pytest-django == 4.2.0
model-bakery == 1.3.1
